# Move relative time calculations to tracer instance