--- a/dd-trace-api/src/main/java/datadog/trace/api/ConfigDefaults.java
+++ b/dd-trace-api/src/main/java/datadog/trace/api/ConfigDefaults.java
@@ -51,8 +51,7 @@ public final class ConfigDefaults {
   static final boolean DEFAULT_JMX_FETCH_ENABLED = true;
   static final boolean DEFAULT_TRACE_AGENT_V05_ENABLED = false;
 
-  static final int DEFAULT_CLOCK_SYNC_PERIOD = 30_000; // milliseconds
-  static final int DEFAULT_CLOCK_DRIFT_LIMIT = 1; // milliseconds
+  static final int DEFAULT_CLOCK_SYNC_PERIOD = 30; // seconds
 
   static final boolean DEFAULT_JMX_FETCH_MULTIPLE_RUNTIME_SERVICES_ENABLED = false;
   static final int DEFAULT_JMX_FETCH_MULTIPLE_RUNTIME_SERVICES_LIMIT = 10;