--- a/dd-trace-core/src/main/java/datadog/trace/core/DDSpan.java
+++ b/dd-trace-core/src/main/java/datadog/trace/core/DDSpan.java
@@ -114,7 +114,7 @@ public class DDSpan implements AgentSpan, CoreSpan<DDSpan>, AttachableWrapper {
     } else {
       startTimeNano = MICROSECONDS.toNanos(timestampMicro);
       externalClock = true;
-      context.getTrace().touch(); // Update lastReferenced
+      context.getTrace().touch(); // external clock: explicitly update lastReferenced
     }
   }
 
@@ -157,8 +157,8 @@ public class DDSpan implements AgentSpan, CoreSpan<DDSpan>, AttachableWrapper {
       durationNano += MICROSECONDS.toNanos(externalOffsetMicros);
     } else {
       durationNano = MICROSECONDS.toNanos(stopTimeMicros) - startTimeNano;
+      context.getTrace().touch(); // external clock: explicitly update lastReferenced
     }
-    context.getTrace().touch(); // Update timestamp
     finishAndAddToTrace(durationNano);
   }