With this change we call `System.currentTimeMillis()` once when creating the tracer - we then use `System.nanoTime()` for all relative times of traces under that tracer, plus a periodic adjustment. As well as reducing the CPU overhead of each trace it also saves us 16 bytes in `PendingTrace` because we no longer need to save the start time and start ticks per-trace.

Every 30 seconds (configurable) we check for clock drift/skew by calling `System.currentTimeMillis()` and comparing that to our calculated time from the initial tracer wall clock time. Any difference more than 1ms is taken from future trace times while keeping the sub-nanosecond time intact.

Note: trace tests that don't involve synthetic spans now sort their test traces by the start time of the local root span (before we used the creation time of the `PendingTrace` structure which was before any spans were created.)

"time-in-queue" messaging tests are slightly different because the start time of the synthetic "time-in-queue" span is explicitly set back to the time the message was produced (with milliseconds granularity) and this isn't stable enough for sorting - so for these tests we sort by span id instead.

Simple span benchmark...

before: 580.509 ±(99.9%) 3.942 ns/op

<img width="1440" alt="Screen Shot 2021-12-21 at 01 27 05" src="https://user-images.githubusercontent.com/145851/146859694-06e37629-c8c7-4801-982c-b6fe4fa78586.png">

after: 501.727 ±(99.9%) 2.593 ns/op

<img width="1440" alt="Screen Shot 2021-12-21 at 01 27 32" src="https://user-images.githubusercontent.com/145851/146859708-98a93076-bdce-43a2-b55d-5f07d7a874be.png">