@luhenry 2021-12-13
> I'd be interested to hear views on this change - especially whether there could be issues for long-lived applications.

What would be your concern here? For any matching of profiling-tracing, because we rely on JFR events, the clock is going to be the same for all profiling events. That might become a problem if we end up not using JFR events for matching, but we aren't there just yet.