@mcculls 2021-12-13
@luhenry mainly I wanted to confirm that it's safe to just rely on `System.nanoTime()` once we've taken that first baseline time+ticks. I think so, but wanted to check with everyone in case there was a subtle reason the original code didn't do that.