@mcculls 2021-12-13
@bantonsson`System.nanoTime()` isn't affected by system clock changes, so relative times are fine - but I could see that if the original wall clock was off (and subsequently corrected) then we wouldn't adjust the start times we reported after the change (they'd be "correct" relative to that initial time, but not the updated time.)