@jbachorik 2021-12-13
JFR is using similar technique. But the difference is that it will recapture the epoch millis per each chunk - thus solving the problem for long lived applications as the time sources will be synchronised regularly.