@bantonsson 2021-12-13
@mcculls My point exactly. I'm not worried about relative times within a Trace, but the fact that start times could drift over time. The old solution compensated for every Trace start.