@mcculls 2021-12-13
@jbachorik interesting, I wonder if we could do something similar - ie. still compensate over time, but not on every trace