@bantonsson 2021-12-13
@mcculls We could definitely compensate over time, i.e. if the nano time diff is greater than X, on every X pending trace, or something similar.